bug fix: comments modified in round trip
If a comment script step with new lines in it was copied and converted to
text, then copied and converted back to FM, it would display as multi-line
in the Script Workspace (like it does on Mac).

While it would be great if FM on Windows displayed comments the same as on
Mac, I am choosing to convert back in the original format.